refactor(#137/*): auth success code global/response 로 이동
